+++ tests/CSharpLanguageServer.Tests/CompletionTests.fs
@@ -24,8 +24,6 @@ let testCompletionWorks () =
     let completion0: U2<CompletionItem array, CompletionList> option =
         client.Request("textDocument/completion", completionParams0)
 
-    let mutable completionItemForMethodA: option<CompletionItem> = None
-
     match completion0 with
     | Some (U2.C2 cl) ->
         Assert.IsTrue(cl.IsIncomplete)
@@ -34,30 +32,56 @@ let testCompletionWorks () =
 
         let methodAItem = cl.Items |> Seq.tryFind (fun i -> i.Label = "MethodA")
         match methodAItem with
+        | None -> failwith "an item with Label 'MethodA' was expected for completion at this position"
         | Some item ->
-            completionItemForMethodA <- Some item
-            Assert.IsFalse(item.LabelDetails.IsSome)
+            Assert.AreEqual(item.Label, "MethodA")
+            Assert.IsFalse(item.Detail.IsSome)
             Assert.IsFalse(item.Documentation.IsSome)
             Assert.IsFalse(item.Tags.IsSome)
-            Assert.IsFalse(item.Detail.IsSome)
+            Assert.AreEqual(item.InsertText, Some "MethodA")
             Assert.AreEqual(Some CompletionItemKind.Method, item.Kind)
             Assert.AreEqual(Some "MethodA", item.SortText)
             Assert.AreEqual(Some "MethodA", item.FilterText)
-            Assert.AreEqual(Some InsertTextFormat.PlainText, item.InsertTextFormat)
+            Assert.AreEqual(None, item.InsertTextFormat)
             Assert.IsFalse(item.CommitCharacters.IsSome)
             Assert.IsFalse(item.TextEdit.IsSome)
             Assert.IsTrue(item.Data.IsSome)
+
+            let itemResolved: CompletionItem = client.Request("completionItem/resolve", item)
+
+            Assert.AreEqual(itemResolved.Detail, Some "void Class.MethodA(string arg)")
+            Assert.IsFalse(itemResolved.Documentation.IsSome)
+
+        let getHashCodeItem = cl.Items |> Seq.tryFind (fun i -> i.Label = "GetHashCode")
+        match getHashCodeItem with
+        | None -> failwith "an item with Label 'GetHashCode' was expected for completion at this position"
+        | Some item ->
+            Assert.AreEqual(item.Label, "GetHashCode")
+            Assert.IsFalse(item.Detail.IsSome)
+            Assert.IsFalse(item.Documentation.IsSome)
+            Assert.IsFalse(item.Tags.IsSome)
+            Assert.AreEqual(item.InsertText, Some "GetHashCode")
+            Assert.AreEqual(Some CompletionItemKind.Method, item.Kind)
+            Assert.AreEqual(Some "GetHashCode", item.SortText)
+            Assert.AreEqual(Some "GetHashCode", item.FilterText)
+            Assert.AreEqual(None, item.InsertTextFormat)
+            Assert.IsFalse(item.CommitCharacters.IsSome)
+            Assert.IsFalse(item.TextEdit.IsSome)
+            Assert.IsTrue(item.Data.IsSome)
+
+            let itemResolved: CompletionItem = client.Request("completionItem/resolve", item)
+
+            Assert.AreEqual(itemResolved.Detail, Some "int object.GetHashCode()")
+            Assert.IsTrue(itemResolved.Documentation.IsSome)
+
+            match itemResolved.Documentation with
+            | Some (U2.C2 markup) ->
+                Assert.AreEqual(MarkupKind.PlainText, markup.Kind)
+                Assert.AreEqual("Serves as the default hash function.", markup.Value)
+            | _ -> failwith "Documentation w/ Kind=Markdown was expected for GetHashCode"
             ()
 
-        | _ -> failwith "an item with Label 'MethodA' was expected for completion at this position"
-
     | _ -> failwith "Some U2.C1 was expected"
-
-    // now check we can resolve completion item data
-    let resolvedItem : CompletionItem =
-        client.Request("completionItem/resolve", completionItemForMethodA.Value)
-
-    Assert.IsTrue(resolvedItem.Documentation.IsSome)
     ()
 
 [<TestCase>]
@@ -85,10 +109,15 @@ let testCompletionWorksForExtensionMethods () =
 
         let methodBItem = cl.Items |> Seq.tryFind (fun i -> i.Label = "MethodB")
         match methodBItem with
+        | None -> failwith "an item with Label 'MethodB' was expected for completion at this position"
         | Some item ->
+            Assert.AreEqual("MethodB", item.Label)
+            Assert.IsFalse(item.Detail.IsSome)
             Assert.AreEqual(Some CompletionItemKind.Method, item.Kind)
-            ()
 
-        | _ -> failwith "an item with Label 'MethodB' was expected for completion at this position"
+            let itemResolved: CompletionItem = client.Request("completionItem/resolve", item)
+
+            Assert.AreEqual(itemResolved.Detail, Some "(extension) string Class.MethodB()")
+            Assert.IsFalse(itemResolved.Documentation.IsSome)
 
     | _ -> failwith "Some U2.C1 was expected"